+++ R/mrs_read_ima.R
@@ -61,7 +61,7 @@ read_ima <- function(fname, verbose = FALSE) {
 
 #' Read a directory containing Siemens MRS IMA files and combine along the coil
 #' dimension. Note that the coil ID is inferred from the sorted file name and
-#' should be checked when consistancy is requried between two directories.
+#' should be checked when consistency is required between two directories.
 #' @param dir data directory path.
 #' @return mrs_data object.
 #' @export