+++ NEWS.md
@@ -1,5 +1,5 @@
 # spant 0.16.0
-* Default rats method improved for multidimensional datasets.
+* Default rats method improved to work with multidimensional datasets.
 * Added norm_mrs function to normalise the intenisty of spectral data.
 * Added bc_constant function to correct spectral baselines by a constant offset.
 * Added re_weighting function to apply a resolution enhancement weighting to the