+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added bc_constant function to correct spectral baselines by a constant offset.
 * Added re_weighting function to apply a resolution enhancement weighting to the
 FID.
+* Performance improvement for apodise_xy function.
 
 # spant 0.15.0
 * Bug fix for using auto_phase function with a single spectrum.