+++ NEWS.md
@@ -9,6 +9,7 @@ paper).
 FID.
 * Performance improvement for apodise_xy function.
 * sd function now works for mrs_data.
+* Added 2D MRSI support for Siemens IMA format.
 
 # spant 0.15.0
 * Bug fix for using auto_phase function with a single spectrum.